Adding proper message length
Collects all the things that need to be sent and puts them in a buffer, then adds the size of that buffer to the buffer itself before sending. The length variable may be unneeded as you could probably just do sizeof(send_buffer), but this is how it is so yea.